Refactor coincall ticker streams and update timestamp handling
- Refactored `coincall_ticker_streams.rs` to dynamically generate BTC options subscriptions.
- Updated timestamp fields from `i128` to `i64` across Coincall authentication and request handling for consistency.
- Exposed `quicksort_iterative` and `partition` functions in `subscription/book.rs` for external use.
- Add missing imports in example comments

+++ src/exchange/coincall/option/ticker.rs
@@ -427,8 +427,8 @@ fn validate_and_format_instrument_name(instrument: &Instrument) -> Result<String
 fn generate_secret(
     auth: &CoincallOptionAuthData,
     instrument_name: &str,
-    ts: i128,
-    tsdiff: i128,
+    ts: i64,
+    tsdiff: i64,
 ) -> Result<String, DataError> {
     let sig_params = SigGenParams {
         key: auth.key.clone(),
@@ -452,8 +452,8 @@ async fn send_request(
     auth: &CoincallOptionAuthData,
     instrument_name: &str,
     sign: &str,
-    ts: i128,
-    tsdiff: i128,
+    ts: i64,
+    tsdiff: i64,
 ) -> Result<reqwest::Response, DataError> {
     let req = reqwest::Client::new()
         .get(&format!(
